Refactor numberOfWins parameter Change type from int to long

+++ src/main/java/dev/bibikvlad/mastermind/model/leaderboard/WinsLeaderboardEntry.java
@@ -1,4 +1,4 @@
 package dev.bibikvlad.mastermind.model.leaderboard;
 
-public record WinsLeaderboardEntry(String playerName, int numberOfWins) {
+public record WinsLeaderboardEntry(String playerName, long numberOfWins) {
 }
